Improvements to Translator
